LayoutTests/imported/w3c/web-platform-tests/webmessaging/postMessage_ArrayBuffer.sub-expected.txt


PASS Test Description: postMessage with ArrayBuffer object.
PASS ArrayBuffer is supported.
